backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-CORS
python-dotenv
requests
psycopg2-binary
python-instagram
boto3
pandas
numpy
pytest
gunicorn
# sp-api
python-amazon-sp-api